ai_engine/requirements.txt
torch 
torchvision
# Utils
fastapi 
uvicorn[standard] 
python-dotenv 
pydantic-settings
onnxruntime
sentence-transformers
playwright
pandas
numpy
requests
pillow
pickle
